# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: GigShield API
  description: Parametric Income Protection for Q-Commerce Delivery Partners
  version: 2.1.0
paths:
...
  /api/admin/dashboard:
    get:
      tags:
      - Admin
      summary: Get Dashboard Kpis
      description: Get all 8 dashboard KPIs.
      operationId: get_dashboard_kpis_api_admin_dashboard_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/admin/claims/queue:
    get:
      tags:
      - Admin
      summary: Get Claims Queue
      description: Get claims requiring manual review, sorted by fraud tier.
      operationId: get_claims_queue_api_admin_claims_queue_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/admin/analytics:
    get:
      tags:
      - Admin
      summary: Get Analytics
      description: Get analytics data for charts.
      operationId: get_analytics_api_admin_analytics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/admin/payouts:
    get:
      tags:
      - Admin
      summary: Get Payout History
      description: Get all payout transactions.
      operationId: get_payout_history_api_admin_payouts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer